breat animation pipeline to pieces
