+++ nasabot/geo/animation/source/__init__.py
@@ -1,0 +1,3 @@
+from nasabot.geo.animation.source.gibs import GIBSSource
+
+__all__ = [GIBSSource]
