+++ nasabot/geo/animation/timeline/__init__.py
@@ -1,0 +1,3 @@
+from nasabot.geo.animation.timeline.interval import Interval
+
+__all__ = [Interval]
